+++ srcs/print_map.c
@@ -6,14 +6,14 @@
 /*   By: mdeville <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/05 18:21:53 by mdeville          #+#    #+#             */
-/*   Updated: 2018/03/05 18:27:21 by mdeville         ###   ########.fr       */
+/*   Updated: 2018/03/07 15:58:23 by mdeville         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
 #include "ft_printf.h"
 
-void	print_map(int **map, int line_size)
+void	print_map(double **map, int line_size)
 {
 	int i;
 
@@ -27,7 +27,7 @@ void	print_map(int **map, int line_size)
 		i = 0;
 		while (i < line_size)
 		{
-			ft_printf("%d ", (*map)[i]);
+			ft_printf("%d ", (int)((*map)[i]));
 			++i;
 		}
 		write(1, "\n", 1);
